requirements.txt
numpy==2.2.3
pillow==11.1.0
tqdm==4.67.1
